// frontend/src/api/__generated__/models/CommitDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Docsa API
 * Docsa의 백엔드 API 명세입니다.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CommitDto
 */
export interface CommitDto {
    /**
     * 
     * @type {number}
     * @memberof CommitDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof CommitDto
     */
    branchId?: number;
    /**
     * 
     * @type {string}
     * @memberof CommitDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof CommitDto
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof CommitDto
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the CommitDto interface.
 */
export function instanceOfCommitDto(value: object): value is CommitDto {
    return true;
}

export function CommitDtoFromJSON(json: any): CommitDto {
    return CommitDtoFromJSONTyped(json, false);
}

export function CommitDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommitDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'branchId': json['branchId'] == null ? undefined : json['branchId'],
        'title': json['title'] == null ? undefined : json['title'],
        'description': json['description'] == null ? undefined : json['description'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function CommitDtoToJSON(json: any): CommitDto {
    return CommitDtoToJSONTyped(json, false);
}

export function CommitDtoToJSONTyped(value?: CommitDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'branchId': value['branchId'],
        'title': value['title'],
        'description': value['description'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
    };
}