// frontend/src/api/__generated__/models/BranchDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Docsa API
 * Docsa의 백엔드 API 명세입니다.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BranchDto
 */
export interface BranchDto {
    /**
     * 
     * @type {number}
     * @memberof BranchDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BranchDto
     */
    name?: string;
    /**
     * 
     * @type {Date}
     * @memberof BranchDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {number}
     * @memberof BranchDto
     */
    fromCommitId?: number;
    /**
     * 
     * @type {number}
     * @memberof BranchDto
     */
    rootCommitId?: number;
    /**
     * 
     * @type {number}
     * @memberof BranchDto
     */
    leafCommitId?: number;
    /**
     * 
     * @type {number}
     * @memberof BranchDto
     */
    saveId?: number;
}

/**
 * Check if a given object implements the BranchDto interface.
 */
export function instanceOfBranchDto(value: object): value is BranchDto {
    return true;
}

export function BranchDtoFromJSON(json: any): BranchDto {
    return BranchDtoFromJSONTyped(json, false);
}

export function BranchDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BranchDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'fromCommitId': json['fromCommitId'] == null ? undefined : json['fromCommitId'],
        'rootCommitId': json['rootCommitId'] == null ? undefined : json['rootCommitId'],
        'leafCommitId': json['leafCommitId'] == null ? undefined : json['leafCommitId'],
        'saveId': json['saveId'] == null ? undefined : json['saveId'],
    };
}

export function BranchDtoToJSON(json: any): BranchDto {
    return BranchDtoToJSONTyped(json, false);
}

export function BranchDtoToJSONTyped(value?: BranchDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'fromCommitId': value['fromCommitId'],
        'rootCommitId': value['rootCommitId'],
        'leafCommitId': value['leafCommitId'],
        'saveId': value['saveId'],
    };
}