// frontend/src/api/__generated__/models/EdgeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Docsa API
 * Docsa의 백엔드 API 명세입니다.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EdgeDto
 */
export interface EdgeDto {
    /**
     * 
     * @type {number}
     * @memberof EdgeDto
     */
    from?: number;
    /**
     * 
     * @type {number}
     * @memberof EdgeDto
     */
    to?: number;
}

/**
 * Check if a given object implements the EdgeDto interface.
 */
export function instanceOfEdgeDto(value: object): value is EdgeDto {
    return true;
}

export function EdgeDtoFromJSON(json: any): EdgeDto {
    return EdgeDtoFromJSONTyped(json, false);
}

export function EdgeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): EdgeDto {
    if (json == null) {
        return json;
    }
    return {
        
        'from': json['from'] == null ? undefined : json['from'],
        'to': json['to'] == null ? undefined : json['to'],
    };
}

export function EdgeDtoToJSON(json: any): EdgeDto {
    return EdgeDtoToJSONTyped(json, false);
}

export function EdgeDtoToJSONTyped(value?: EdgeDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'from': value['from'],
        'to': value['to'],
    };
}